Rename component files.
